requirements.txt
Flask==3.1.0
tensorflow-cpu==2.16.2
keras==3.13.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.0
gunicorn==21.2.0
requests==2.31.0
Werkzeug==3.1.3
kaggle==1.6.17
Pillow==10.4.0
opencv-python==4.10.0.84